Decouple slub_debug and no_hash_pointers
Set 3 modes to use from the command line: `auto`, `always` and `never`.
Save the policy which can take 3 different values and use a separate boolean
variable to save the final state of the hash pointers, `always` will set the
value to true, `never` will set the value to false. The `auto` setup will
depend on the slub_debug command line selection, if enabled, hash_pointers
will be false, to allow for clear pointers for debug; if not set,
hash_pointers will be set to true.

Signed-off-by: Sergio Perez Gonzalez <sperezglz@gmail.com>