Decouple slub_debug and no_hash_pointers
Set 3 modes to use from the kernel command line: 'auto', 'always' and 'never'.
Internally the hash_pointers is a boolean, 'always' will set the value
to true, 'never' will set the value to false. The 'auto' setup will depend
on compile time selection of CONFIG_SLUB_DEBUG, if '=y', hash_pointers
will be set to false, to allow for clear pointers for debug; if not set,
hash_pointers will be set to true.

Signed-off-by: Sergio Perez Gonzalez <sperezglz@gmail.com>